Fixed an issue with querying weapons
Fixed an issue with querying weapons where it would not work at all
due to the query having inner joins on all relevant tables including
abilities and skills, where a weapon usually has either of these or
none of these. These have been changed to left joins.
Additionally, fixed an issue where the weapon type would not pass
properly, and an unused level parameter was kept in the weapon class