Added limited and removed ability icons
Added the limited icon for abilities on wyrmprints that aren't always
there. For example, facility event effects only appear during facility
events and will be marked as limited by the bot.
Added the removed (X) icon for abilities on wyrmprints that no longer
exist. These are for the older prints that were for raid events and had
"x's bane +y%". Due to changes in the game, these effects no longer
exist for these bosses.
Added the assets for the emoji used for the discord server
Updated the readme for the server updates